hotfix: chức năng validate add import và sửa lại seerder, thêm dữ liệu vào db

+++ database/migrations/2024_09_16_062921_create_receipts_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
             $table->char('supplier_code', 20);
             $table->text('note')->nullable();
             $table->boolean('status')->default(true);
-            $table->integer('receipt_no');
+            $table->bigInteger('receipt_no');
             $table->date('receipt_date')->nullable();
             $table->char('created_by', 20);
             $table->timestamps();
